Renamed formatting.js to formattingService.js. [Just for consistency]
